+++ packages/akeru-frontend/app/components/sections/hero-section.tsx
@@ -1,0 +1,15 @@
+function HeroSection() {
+  return (
+    <section className="text-center max-w-xl mt-12 mx-auto md:mt-40">
+      <h2 className="font-montserrat font-bold text-5xl capitalize">
+        The best API to setup your AI Project{" "}
+      </h2>
+      <p className="font-montserrat mt-3 text-base font-medium opacity-70">
+        Lorem ipsum dolor sit amet consectetur. Turpis pellentesque praesent
+        mattis a nisl augue. Ipsum nullam at facilisis tortor. Feugiat{" "}
+      </p>
+    </section>
+  );
+}
+
+export default HeroSection;
