+++ packages/akeru-frontend/app/components/ui/form/Form.tsx
@@ -1,32 +1,27 @@
-import { Dispatch, SetStateAction } from "react";
 import FormInput from "./FormInput";
-import { IoMdClose } from "react-icons/io";
 
 type FormProps = {
   formAction: (payload: FormData) => void;
-  setShowForm: Dispatch<SetStateAction<boolean>>;
 };
 
-function Form({ formAction, setShowForm }: FormProps) {
+function Form({ formAction }: FormProps) {
   return (
     <form
-      className="w-full  px-6 rounded-xl h-fit pb-10 pt-5 flex flex-col gap-2.5"
+      className="border border-white border-opacity-30 backdrop-blur-lg w-full px-6 md:px-16 rounded-xl h-fit flex flex-col gap-4 bg-white max-w-2xl mx-auto py-10 bg-opacity-20 shadow-2xl shadow-green-900"
       action={formAction}
     >
-      <span className=" text-center text-opacity-90 sm:text-sm">
+      <span className="text-green-50 text-opacity-80 text-sm md:text-base">
         Stay in the loop! Enter your email address to receive an exclusive
         notification when Akeru goes live.
       </span>
-      <div className="tablet_max:mt-2 mobile_lg:mt-3.5">
-        <FormInput
-          type="email"
-          name="emailAddress"
-          placeholder="Email Address"
-          required
-        />
-      </div>
-      <div className="tablet_max:mt-2 mobile_lg:mt-3.5">
-        <button className="bg-green-900 w-full py-3 rounded-md duration-200 hover:opacity-80">
+      <FormInput
+        type="email"
+        name="emailAddress"
+        placeholder="Email Address"
+        required
+      />
+      <div className="">
+        <button className="bg-gradient-to-b text-sm md:text-base from-green-800 to-green-900 shadow-md w-full py-3 rounded-md duration-200 hover:shadow-lg">
           Get Akeru
         </button>
       </div>
